Better support plurals
If CDS tries to serve a translation of a string that is saved in
Transifex as pluralized, it will not have a way of knowing the main
variable's name. Instead it will use '???'.

So, if the source string in the native application is a compatible
pluralized string and CDS returned a pluralized translation with '???'
as its variable name, this change replaces '???' with the variable name
captured from the source string before serving the translation back to
the application.

A compatible pluralized string is one with the plural pattern defined at
the outermost layer. So for example:

    {cnt, plural, one {You have # message} other {You have # messages}}

is a compatible plrualized string, while:

    You have {cnt, plural, one {# message} other {# messages}}

is not.