Fix templatetag when rendering missing translations
Here is the problem:

When there is an untranslated string in the production database, the
cache will eventually have an entry like `key: {string: ""}`. This gets
handled by the missing policy properly.

When, however, the string is totally unknown to the production database,
the templatetag would render "None". This was done because at some point
we call `escape_html(translation_icu_template)` on a `None` value, when
`escape_html` really wants to return a string.