Allow '=' in first arg of tag if string literal
Previously, we checked for the existense of a '=' on the first argument
of the t-tag to determine if it is a parameter, in which case, we
assumed the user was using the block-style syntax of the t-tag. This
obviously isn't the case if the first argument is a string literal that
has the '=' within it.