Allow Native to start without syncing translations
Introduce a new setting, `SKIP_TRANSLATIONS_SYNC`; if set to True,
no translations will be automatically fetch from CDS and the syncing
daemon will not start.

Introduce a way to have the initial syncing of translations but
not start the daemon, by setting `TRANSIFEX_SYNC_INTERVAL=0`.

Fix exception when daemon is not started. Now that the SDK allows the
syncing daemon to never start, an exception occurred because
`Thread.is_alive()` returned False, which triggered `Thread.isAlive`
to be called, resulting to an exception since this was removed
after a certain Python version.

Create dedicated module to host the parsing of some Django settings
now that cache support is also added alongside missing and error
policies.