Improve t-tag performance
The problem is that for every t-tag rendering, we go over the **entire**
template context and conditionally HTML-escape it before feeding it to
python-native to render the ICU template. If the context is big, this
can slow things down a lot.

This commit attempts to go over the ICU template to find all parameter
names that it will need. Then it only searches these names against the
context to conditionally escape. This should make things better.