Support custom keys in Python & Django
Allow users to optionally define custom keys for strings
in Python and Django.

Developers can use the special `_key` parameter in order to set
a custom key, instead of using the one automatically generated
by the source string and the context.

Amend the `--verbose` option of the push command so that it always
prints the key of each string.