Add fallback value
Django's makemigrations command has some trouble with our LazyStrings.
When LazyStrings are used as choices in a `models.TextChoices` subclass,
the migration file will either have:

1. The return value of `__repr__`
2. The underlying string, which in the initial implementation was ''

We add an optional fallback_value keyword argument to be used as both
the value of the underlying string and as the return value of `__repr__`
so that makemigrations will produce a correct migration file.