Refactor transifex.api's exceptions for down/uploads
Previously, for uploads and downloads we could get a "successful"
response that includes an error report for a download or upload. This
means that as far as {json:api} semantics go, no error has happened.
However, we intercepted this response and raised it as if it was a
{json:api} exception.

This raising was buggy because it didn't use the `JsonApiException.new`
invocation (and thus can't be captured with `JsonApiException.get`) and
also it didn't have the third required argument which was the response.

Instead of fixing the aforementioned bugs however, we decided to
implement new exception classes in `transifex.api` (which is the SDK
itself, as opposed to `transifex.api.jsonapi` which is the "SDK
library"). This makes it clearer to the user that as far as the
{json:api} semantics are concerned, no error has occured. The problem is
that the download or upload "job"'s status (which was successfully
retrieved) is reporting an error.

The new exception classes give priority to their first argument which is
the 'detail' field of the first error included in the response. The rest
of the errors, if any, will be accessible from the exception's second
argument. Keep in mind that at the time of writing, Transifex APIv3 will
only include one error in the case of an unsuccessful upload or
download.